Get card by set code and collector number

Get card by set code and collector number

get_card_by_code(
  code = "wwk",
  number = 31,
  lang = "en",
  format = "json",
  face = "front",
  version = "large",
  pretty = FALSE,
  rulings = FALSE
)

Arguments

code
the 3 to 5 letter set code

number
the collector number of the card (must be whole number)

lang
the 2 to 3 letter language code as described at https://scryfall.com/docs/api/languages

format
The format to return the data in. "json", "CSV", "image"

face
The face that should be returned if format selected is "image"

version
The size of the image to return when using "image"

pretty
Should the JSON be prettified

rulings
(Boolean) should the rulings for the card be returned?