Get card by ID

Get a single card by it's given id from one of the following ID schemes:

multiverse

mtgo

arena

tcgplayer

cardmarket

scryfall See details for information about each of these ID systems.

get_card_by_id(
  id = 195297,
  format = "json",
  face = "front",
  version = "large",
  pretty = FALSE,
  id_type = "multiverse",
  rulings = FALSE
)

Arguments

id
Arena ID

format
The format to return the data in. "json", "text", "image"

face
The face that should be returned if format selected is "image"

version
The size of the image to return when using "image"

pretty
Should the JSON be prettified

id_type
what kind of id is being provided See details for options.

rulings
Boolean should the rulings for the card be returned?

Value

either a list, data frame, or image depending on format

Details

id_type

'multiverse': Get a single card by given Multiverse ID. If the card has multiple Multiverse IDs this method finds both of them. A Multiverse ID is the unique identifier from Wizards of the Coast's Gatherer database.

`mtgo': Get a single card by given it's MTGO ID, also known as Catalog ID. These can be either the mtgo_id or it's mtgo_foil_id.

'arena': Get a single card given it's arena ID

'tcgplayer:' Get a single card with the given tcgplayer_id, also known as the productId on TCGplayer’s API

'cardmarket' Get a single card with the given cardmarket_id, also known as the idProduct" or the Product ID on Cardmarket’s APIs

'scryfall' Get a single card with the given Scryfall ID