Search for cards on Scryfall

Get a list of cards found using a fulltext search string. This string uses the same syntax as Scryfall's site This is pagenated, returning a maximum of 175 cards at a time

get_cards_by_search(
  q = "*",
  unique = "cards",
  order = "name",
  dir = "auto",
  include_extras = FALSE,
  include_variations = FALSE,
  page = 1,
  format = "json",
  pretty = FALSE
)

Arguments

q
the string query using Scryfall Syntax

unique
the rollup mode for removing duplicates. See details

order
the method to sort returned cards. See details

dir
direction to sort cards

include_extras
Should the search include extra cards (tokens, planes, vanguards, etc)

include_variations
Should rare card varients be included Defaults to false

page
the page to return defaults to 1

format
The format to return the data in. json or CSV

pretty
Should the JSON be prettified

Value

A list

Details

unique This parameter specifies if Scryfall should remove duplicate results in your query. The options are:

"cards" (default): Removes duplicate gameplay objects (cards that share a name and have the same functionality).

"art" : Returns only one copy of each unique artwork for matching cards.

"prints": Returns all prints for all cards matched (disables rollup).

order This parameter determines how Scryfall should sort the returned cards.

"name" (default): by card name

"set": by their set and collector number

"released": by release date

"rarity": by card Rarity

"usd": by card lowest known US dollar price

"tix": by lowest known TIX price

"eur": by lowest known Euro price

"cmc": by converted mana cost

"power": by power

"toughness": by toughness

"edhrec": by edhrec ranking

"artist" by front side artist name